wd-timetracking-contract/src/main/java/de/webducer/android/zeiterfassung/contract/enums/ITranslatableEnum.java
/*
 * The MIT License (MIT)
 *
 * Copyright (c) 2015 Eugen [WebDucer] Richter
 *
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
 * in the Software without restriction, including without limitation the rights
 * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 * copies of the Software, and to permit persons to whom the Software is
 * furnished to do so, subject to the following conditions:
 *
 * The above copyright notice and this permission notice shall be included in all
 * copies or substantial portions of the Software.
 *
 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
 * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 * SOFTWARE.
 */

package de.webducer.android.zeiterfassung.contract.enums;

import android.content.Context;

/**
 * Interface for enums with translation resource
 *
 * @author WebDucer - IT &amp; Internet Service
 * @version 0.2
 * @since 2015-04-04
 */
public interface ITranslatableEnum {
   /**
    * Get int representation of the value (eg. for DB)
    *
    * @return Int representation of enum value
    */
   public int getEnumCode();

   /**
    * Get translation id for the enum value
    *
    * @return Translation key for further translation
    */
   public int getTranslationId();

   /**
    * Get translated value of the enum value
    *
    * @param context App context
    * @return Translated enum value
    */
   public String getTranslatedValue(Context context);

   /**
    * Get Enum value by given enum code
    *
    * @param enumCode Enum code (eg. from data base)
    * @return Resolved enum value or 'None' if no match
    */
   public ITranslatableEnum getByEnumCode(int enumCode);
}
